Str concat 3

+++ 0x0B-malloc_free/2-str_concat.c
@@ -16,7 +16,7 @@ char *str_concat(char *s1, char *s2)
 	if (s2 == NULL)
 		s2 = "";
 
-	long int size = strlen(s1) + strlen(s2) + 1;
+	int size = strlen(s1) + strlen(s2) + 1;
 	char *ptr = malloc(sizeof(char) * size);
 
 	if (ptr == NULL)
